Move build endpoint handler from daemon
